Refactor different runtimes into its own folder/module
