+++ elyra/metadata/schemas/kfp.json
@@ -4,7 +4,7 @@
   "name": "kfp",
   "display_name": "Kubeflow Pipelines",
   "namespace": "runtimes",
-  "schema_filter_class_name": "elyra.pipeline.processor_kfp.KfpSchemaFilter",
+  "schema_filter_class_name": "elyra.pipeline.kfp.processor_kfp.KfpSchemaFilter",
   "uihints": {
     "title": "Kubeflow Pipelines runtimes",
     "icon": "elyra:runtimes",
