+++ elyra/pipeline/airflow/tests/test_component_parser_airflow.py
@@ -20,9 +20,9 @@ import jupyter_core.paths
 
 from elyra.metadata.manager import MetadataManager
 from elyra.metadata.metadata import Metadata
+from elyra.pipeline.airflow.component_parser_airflow import AirflowComponentParser
 from elyra.pipeline.component import FilesystemComponentReader
 from elyra.pipeline.component import UrlComponentReader
-from elyra.pipeline.component_parser_airflow import AirflowComponentParser
 from elyra.pipeline.component_registry import ComponentRegistry
 
 COMPONENT_CATALOG_DIRECTORY = os.path.join(jupyter_core.paths.ENV_JUPYTER_PATH[0], 'components')
@@ -30,7 +30,8 @@ COMPONENT_CATALOG_DIRECTORY = os.path.join(jupyter_core.paths.ENV_JUPYTER_PATH[0
 
 def _get_resource_path(filename):
     root = os.path.realpath(os.path.join(os.getcwd(), os.path.dirname(__file__)))
-    resource_path = os.path.join(root, 'resources', 'components', filename)
+    resource_path = os.path.join(root, '../../tests', 'resources', 'components', filename)
+    resource_path = os.path.normpath(resource_path)
 
     return resource_path
 
@@ -52,8 +53,8 @@ def test_modify_component_registries():
     metadata_manager = MetadataManager(namespace=MetadataManager.NAMESPACE_COMPONENT_REGISTRIES)
 
     # Create new registry instance with a single URL-based component
-    paths = ["https://raw.githubusercontent.com/elyra-ai/elyra/master/elyra/pipeline/tests/resources/components"
-             "/airflow_test_operator.py"]
+    paths = ["https://raw.githubusercontent.com/elyra-ai/elyra/master/elyra/pipeline/tests/resources/components/"
+             "airflow_test_operator.py"]
 
     instance_metadata = {
         "description": "A test registry",
